Structural changes to the react application. Splitting functionality into seperate components. Working on a rudimentary text box in the bottom of the screen. The idea is to have something similiar to a command line.
